+++ kasa/exceptions.py
@@ -126,50 +126,6 @@ class SmartErrorCode(IntEnum):
     DST_ERROR = -2301
     DST_SAVE_ERROR = -2302
 
-    # Camera error codes
-    SESSION_EXPIRED = -40401
-    HOMEKIT_LOGIN_FAIL = -40412
-    DEVICE_BLOCKED = -40404
-    DEVICE_FACTORY = -40405
-    OUT_OF_LIMIT = -40406
-    OTHER_ERROR = -40407
-    SYSTEM_BLOCKED = -40408
-    NONCE_EXPIRED = -40409
-    FFS_NONE_PWD = -90000
-    TIMEOUT_ERROR = 40108
-    UNSUPPORTED_METHOD = -40106
-    ONE_SECOND_REPEAT_REQUEST = -40109
-    INVALID_NONCE = -40413
-    PROTOCOL_FORMAT_ERROR = -40210
-    IP_CONFLICT = -40321
-    DIAGNOSE_TYPE_NOT_SUPPORT = -69051
-    DIAGNOSE_TASK_FULL = -69052
-    DIAGNOSE_TASK_BUSY = -69053
-    DIAGNOSE_INTERNAL_ERROR = -69055
-    DIAGNOSE_ID_NOT_FOUND = -69056
-    DIAGNOSE_TASK_NULL = -69057
-    CLOUD_LINK_DOWN = -69060
-    ONVIF_SET_WRONG_TIME = -69061
-    CLOUD_NTP_NO_RESPONSE = -69062
-    CLOUD_GET_WRONG_TIME = -69063
-    SNTP_SRV_NO_RESPONSE = -69064
-    SNTP_GET_WRONG_TIME = -69065
-    LINK_UNCONNECTED = -69076
-    WIFI_SIGNAL_WEAK = -69077
-    LOCAL_NETWORK_POOR = -69078
-    CLOUD_NETWORK_POOR = -69079
-    INTER_NETWORK_POOR = -69080
-    DNS_TIMEOUT = -69081
-    DNS_ERROR = -69082
-    PING_NO_RESPONSE = -69083
-    DHCP_MULTI_SERVER = -69084
-    DHCP_ERROR = -69085
-    STREAM_SESSION_CLOSE = -69094
-    STREAM_BITRATE_EXCEPTION = -69095
-    STREAM_FULL = -69096
-    STREAM_NO_INTERNET = -69097
-    HARDWIRED_NOT_FOUND = -72101
-
     # Library internal for unknown error codes
     INTERNAL_UNKNOWN_ERROR = -100_000
     # Library internal for query errors
