+++ kasa/experimental/sslaestransport.py
@@ -1,8 +1,4 @@
-"""Implementation of the TP-Link AES transport.
-
-Based on the work of https://github.com/petretiandrea/plugp100
-under compatible GNU GPL3 license.
-"""
+"""Implementation of the TP-Link SSL AES transport."""
 
 from __future__ import annotations
 
@@ -12,30 +8,26 @@ import logging
 import secrets
 import ssl
 import time
-from enum import Enum, auto
+from enum import Enum, IntEnum, auto
+from functools import cache
 from typing import TYPE_CHECKING, Any, Dict, cast
 
-from cryptography.hazmat.primitives import padding
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
 from urllib3.util import create_urllib3_context
 from yarl import URL
 
-from .aestransport import AesEncyptionSession
-from .credentials import Credentials
-from .deviceconfig import DeviceConfig
-from .exceptions import (
-    SMART_AUTHENTICATION_ERRORS,
-    SMART_RETRYABLE_ERRORS,
+from ..aestransport import AesEncyptionSession
+from ..credentials import Credentials
+from ..deviceconfig import DeviceConfig
+from ..exceptions import (
     AuthenticationError,
     DeviceError,
     KasaException,
-    SmartErrorCode,
     _RetryableError,
 )
-from .httpclient import HttpClient
-from .json import dumps as json_dumps
-from .json import loads as json_loads
-from .protocol import BaseTransport
+from ..httpclient import HttpClient
+from ..json import dumps as json_dumps
+from ..json import loads as json_loads
+from ..protocol import BaseTransport
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -129,6 +121,7 @@ class SslAesTransport(BaseTransport):
         }
         self._seq: int | None = None
         self._pwd_hash: str | None = None
+        self._username: str | None = None
         if self._credentials != Credentials() and self._credentials:
             self._username = self._credentials.username
         elif self._credentials_hash:
@@ -425,25 +418,77 @@ class SslAesTransport(BaseTransport):
         self._local_nonce = None
 
 
-class SslAesEncyptionSession:
-    """Class for an AES encryption session."""
+class SmartErrorCode(IntEnum):
+    """Smart error codes for this transport."""
 
-    def __init__(self, key, iv):
-        self.cipher = Cipher(algorithms.AES(key), modes.CBC(iv))
-        self.padding_strategy = padding.PKCS7(algorithms.AES.block_size)
+    def __str__(self):
+        return f"{self.name}({self.value})"
 
-    def encrypt(self, data) -> bytes:
-        """Encrypt the message."""
-        encryptor = self.cipher.encryptor()
-        padder = self.padding_strategy.padder()
-        padded_data = padder.update(data) + padder.finalize()
-        encrypted = encryptor.update(padded_data) + encryptor.finalize()
-        return base64.b64encode(encrypted)
+    @staticmethod
+    @cache
+    def from_int(value: int) -> SmartErrorCode:
+        """Convert an integer to a SmartErrorCode."""
+        return SmartErrorCode(value)
 
-    def decrypt(self, data) -> str:
-        """Decrypt the message."""
-        decryptor = self.cipher.decryptor()
-        unpadder = self.padding_strategy.unpadder()
-        decrypted = decryptor.update(base64.b64decode(data)) + decryptor.finalize()
-        unpadded_data = unpadder.update(decrypted) + unpadder.finalize()
-        return unpadded_data.decode()
+    SUCCESS = 0
+
+    SYSTEM_ERROR = -40101
+    INVALID_ARGUMENTS = -40209
+
+    # Camera error codes
+    SESSION_EXPIRED = -40401
+    HOMEKIT_LOGIN_FAIL = -40412
+    DEVICE_BLOCKED = -40404
+    DEVICE_FACTORY = -40405
+    OUT_OF_LIMIT = -40406
+    OTHER_ERROR = -40407
+    SYSTEM_BLOCKED = -40408
+    NONCE_EXPIRED = -40409
+    FFS_NONE_PWD = -90000
+    TIMEOUT_ERROR = 40108
+    UNSUPPORTED_METHOD = -40106
+    ONE_SECOND_REPEAT_REQUEST = -40109
+    INVALID_NONCE = -40413
+    PROTOCOL_FORMAT_ERROR = -40210
+    IP_CONFLICT = -40321
+    DIAGNOSE_TYPE_NOT_SUPPORT = -69051
+    DIAGNOSE_TASK_FULL = -69052
+    DIAGNOSE_TASK_BUSY = -69053
+    DIAGNOSE_INTERNAL_ERROR = -69055
+    DIAGNOSE_ID_NOT_FOUND = -69056
+    DIAGNOSE_TASK_NULL = -69057
+    CLOUD_LINK_DOWN = -69060
+    ONVIF_SET_WRONG_TIME = -69061
+    CLOUD_NTP_NO_RESPONSE = -69062
+    CLOUD_GET_WRONG_TIME = -69063
+    SNTP_SRV_NO_RESPONSE = -69064
+    SNTP_GET_WRONG_TIME = -69065
+    LINK_UNCONNECTED = -69076
+    WIFI_SIGNAL_WEAK = -69077
+    LOCAL_NETWORK_POOR = -69078
+    CLOUD_NETWORK_POOR = -69079
+    INTER_NETWORK_POOR = -69080
+    DNS_TIMEOUT = -69081
+    DNS_ERROR = -69082
+    PING_NO_RESPONSE = -69083
+    DHCP_MULTI_SERVER = -69084
+    DHCP_ERROR = -69085
+    STREAM_SESSION_CLOSE = -69094
+    STREAM_BITRATE_EXCEPTION = -69095
+    STREAM_FULL = -69096
+    STREAM_NO_INTERNET = -69097
+    HARDWIRED_NOT_FOUND = -72101
+
+    # Library internal for unknown error codes
+    INTERNAL_UNKNOWN_ERROR = -100_000
+    # Library internal for query errors
+    INTERNAL_QUERY_ERROR = -100_001
+
+
+SMART_RETRYABLE_ERRORS = [
+    SmartErrorCode.SESSION_EXPIRED,
+]
+
+SMART_AUTHENTICATION_ERRORS = [
+    SmartErrorCode.INVALID_ARGUMENTS,
+]
